+++ src/mnt/ode/services/sqd_file_service.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 import logging
 from typing import TYPE_CHECKING
 
-from mnt.ode.models.layout_model import LayoutModel
+from mnt.ode.models import LayoutModel
 from mnt.pyfiction import read_sqd_layout_100, sqd_parsing_error
 
 if TYPE_CHECKING:
